Rename callon reroll to traitor reroll
